% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_compute.R
\name{A_compute}
\alias{A_compute}
\title{A_compute: Compute adjacency matrix to construct a network}
\usage{
A_compute(df)
}
\arguments{
\item{df}{a data frame containing a column ISRC codes, and a column with the registrant of each ISRC}
}
\value{
returns an adjacency matrix
}
\description{
A_compute: Compute adjacency matrix to construct a network
}
