% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artist_name.R
\name{extract_artist_name}
\alias{extract_artist_name}
\title{Extract artist name}
\usage{
extract_artist_name(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Artist names as a single character string.
}
\description{
Extract artist name
}
