% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_isrc.R
\name{extract_isrc}
\alias{extract_isrc}
\title{Extract ISRC from a data frame}
\usage{
extract_isrc(df)
}
\arguments{
\item{df}{A data frame that may contain the \code{external_ids} column.}
}
\value{
The external_ids column.
}
\description{
Extract ISRC from a data frame
}
