% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_track_attributes.R
\name{extract_track_attributes}
\alias{extract_track_attributes}
\title{Extract attribute data from tracks}
\usage{
extract_track_attributes(track_df)
}
\arguments{
\item{track_df}{Must be a data frame that contains track information, not artists,
playlists or shows.}

\item{df}{A data frame that may contain the \code{external_ids} column.}
}
\value{
data frame with track attributes
}
\description{
Extract attribute data from tracks
}
