% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_track_variables.R
\name{extract_track_variables}
\alias{extract_track_variables}
\title{Extract variable data from tracks}
\usage{
extract_track_variables(track_df, time_query)
}
\arguments{
\item{track_df}{Must be a data frame that contains track information, not artists,
playlists or shows.}

\item{time_query}{the time of the query.}

\item{df}{A data frame that may contain the \code{external_ids} column.}
}
\value{
The external_ids column.
}
\description{
Extract variable data from tracks
}
