% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article_length.R
\name{get_article_length}
\alias{get_article_length}
\title{get_article_length: extract article length (in bytes)}
\usage{
get_article_length(language_code, page_title)
}
\arguments{
\item{language_code}{the language code}

\item{page_title}{the title of the article on wikipedia}
}
\description{
get_article_length: extract article length (in bytes)
}
