% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_edit_time.R
\name{get_average_edit_time}
\alias{get_average_edit_time}
\title{get_average_edit_time: get time between successive edits}
\usage{
get_average_edit_time(language_code, page_title)
}
\arguments{
\item{language_code}{the language code}

\item{page_title}{the title of the article on wikipedia}
}
\description{
get_average_edit_time: get time between successive edits
}
