% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creation_date.R
\name{get_creation_date}
\alias{get_creation_date}
\title{get_creation_date: get the creation date of an article}
\usage{
get_creation_date(language_code, page_title)
}
\arguments{
\item{language_code}{the language code}

\item{page_title}{the title of the article on wikipedia}
}
\description{
get_creation_date: get the creation date of an article
}
