% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number_of_edits.R
\name{get_number_of_edits}
\alias{get_number_of_edits}
\title{get_number_of_edits: get the number of edits of a page}
\usage{
get_number_of_edits(language_code, page_title)
}
\arguments{
\item{language_code}{the language code}

\item{page_title}{the title of the article on wikipedia}
}
\description{
get_number_of_edits: get the number of edits of a page
}
