% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quality_class.R
\name{get_quality_class}
\alias{get_quality_class}
\title{get_quality_class: get page's quality class}
\usage{
get_quality_class(language_code, article_title)
}
\arguments{
\item{language_code}{the language code}

\item{page_title}{the title of the article on wikipedia}
}
\description{
get_quality_class: get page's quality class
}
