% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_count.R
\name{get_reference_count}
\alias{get_reference_count}
\title{get_reference_count: get the number of references in a page}
\usage{
get_reference_count(language_code, page_title)
}
\arguments{
\item{language_code}{the language code}

\item{page_title}{the title of the article on wikipedia}
}
\description{
get_reference_count: get the number of references in a page
}
