% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_editors.R
\name{get_unique_editors}
\alias{get_unique_editors}
\title{get_unique_editors: extract the number of unique editors}
\usage{
get_unique_editors(language_code, page_title)
}
\arguments{
\item{language_code}{the language code}

\item{page_title}{the title of the article on wikipedia}
}
\description{
get_unique_editors: extract the number of unique editors
}
