% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{one_sample}
\alias{one_sample}
\title{one_sample: it returns only one sample using a uniform Metropolist-Hastings Random-Walk algorithm}
\usage{
one_sample(data, d = 0.5, size = 100)
}
\arguments{
\item{data}{a data frame containing ISRCs, Registrant and variable defined in the x parameter}

\item{d}{the probability of selecting a non-neighbour at each step (default is 0.5)}

\item{size}{the desired sample size}
}
\value{
a vector with one sample obtained from the variable specified by x samples and the acceptance rate of the MHRW algorithm
}
\description{
one_sample: it returns only one sample using a uniform Metropolist-Hastings Random-Walk algorithm
}
\examples{

example <- one_sample("popularity", data = sample_data)
example


}
