% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmusesampling.R
\name{openmusesampling}
\alias{openmusesampling}
\title{sampling: function to generate sample mean statistics using a uniform Metropolist-Hastings Random-Walk algorithm}
\usage{
openmusesampling(x, data, B = 1000, d = 0.5, size = 100)
}
\arguments{
\item{x}{a string containing the name of the variable you wish to compute the mean}

\item{data}{a data frame containing ISRCs, Registrant and variable defined in the x parameter}

\item{B}{number of samples to construct the sampling distribution (default is 1000)}

\item{d}{the probability of selecting a non-neighbour at each step (default is 0.5)}

\item{size}{the desired sample size}
}
\value{
a data frame with the sample mean calculated from B samples and the acceptance rate of the MHRW algorithm
}
\description{
sampling: function to generate sample mean statistics using a uniform Metropolist-Hastings Random-Walk algorithm
}
\examples{

example <- openmusesampling("popularity", data = sample_data)
example

The sample mean calculated above should be close to the true sample mean and this true mean should be contained in the 95\% CI
mean(sample_data$popularity)

}
