% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_quantiles.R
\name{permutation_quantiles}
\alias{permutation_quantiles}
\title{permutation_quantiles: this function performs the permutation test for the quantiles of two samples}
\usage{
permutation_quantiles(x1, x2, quantiles = c(0.025, 0.975), N = 10000)
}
\arguments{
\item{x1}{a vector with the first sample}

\item{x2}{a vector with the second sample}

\item{N}{number of resamples to approximate the permutation distribution under the null}
}
\value{
a list containing the permutation distribution, the observed difference in quantiles and the p value
}
\description{
permutation_quantiles: this function performs the permutation test for the quantiles of two samples
}
\examples{

# This function performs the following test: H_0: quantiles are equal against H_1: not equal

N <- 10000
x1 <- rnorm(100)
x2 <- rnorm(100)

permutation_quantiles(x1, x2, N)

}
