% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_sd.R
\name{permutation_sd}
\alias{permutation_sd}
\title{permutation_sd: this function performs the permutation test for the standard deviation of two samples}
\usage{
permutation_sd(x1, x2, N = 10000)
}
\arguments{
\item{x1}{a vector with the first sample}

\item{x2}{a vector with the second sample}

\item{N}{number of resamples to approximate the permutation distribution under the null}
}
\value{
a list containing the permutation distribution, the observed ratio in sd's and the p value
}
\description{
permutation_sd: this function performs the permutation test for the standard deviation of two samples
}
\examples{

# This function performs the following test: H_0: sd's are equal against H_1: not equal

N <- 10000
x1 <- rnorm(100)
x2 <- rnorm(100)

permutation_sd(x1, x2, N)

}
