% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.openmusesampling.R
\name{print.openmusesampling}
\alias{print.openmusesampling}
\title{print.openmusesampling: prints summary statistics of an openmusesampling object}
\usage{
\method{print}{openmusesampling}(x)
}
\arguments{
\item{x}{an object generated by the openmusesampling function}
}
\description{
print.openmusesampling: prints summary statistics of an openmusesampling object
}
