% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_compute.R
\name{Q_compute}
\alias{Q_compute}
\title{Q_compute: constructs the transition matrix Q}
\usage{
Q_compute(A, d = 0.5)
}
\arguments{
\item{A}{an adjacency matrix}

\item{d}{the probability to sample a non conected node in the random walk}
}
\value{
returns the transition matrix of the Markov chain
}
\description{
Q_compute: constructs the transition matrix Q
}
