% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_compute_MH.R
\name{Q_compute_MH}
\alias{Q_compute_MH}
\title{Q_compute_MH: computes the adjusted transition matrix for the Random Walk with jumps}
\usage{
Q_compute_MH(A, Q, d = 0.5)
}
\arguments{
\item{A}{the network adjacency matrix}

\item{Q}{the network transition matrix}

\item{d}{the probability to jump to a non connected node}
}
\value{
returns the adjusted transition matrix
}
\description{
Q_compute_MH: computes the adjusted transition matrix for the Random Walk with jumps
}
