% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Returned data of sample_data}
\format{
A data frame containing 1204 rows and 6 columns
\describe{
  \item{id}{Spotify track IDs}
  \item{isrc}{Track ISRCs}
  \item{registrant}{The code of the registrant that registered the ISRC}
  \item{year}{The year the ISRC was registered}
  \item{designation}{The last five digist of the ISRC}
  \item{popularity}{The popularity of the tracks on Spotify}
}
}
\usage{
sample_data
}
\description{
A data frame containing 1024 ISRCs
}
\details{
This is an example data to show how the Metropolis-Hastings uniform random-walk algorithm works
}
\keyword{datasets}
