% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMH.R
\name{samplerMH}
\alias{samplerMH}
\title{samplerMH a Metropolis-Hastings algorithm to run the Random Walk in the network}
\usage{
samplerMH(A, Q, d = 0.5, size = 100)
}
\arguments{
\item{A}{network adjacency matrix}

\item{Q}{network transition matrix}

\item{d}{the probability to jump to a non connected node}

\item{size}{the size of the sample}
}
\value{
returns a list with random samples from network nodes and the acceptance rate of the jumps
}
\description{
samplerMH a Metropolis-Hastings algorithm to run the Random Walk in the network
}
