% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_tracks_by_term.R
\name{search_tracks_by_term}
\alias{search_tracks_by_term}
\title{search_tracks_by_term: Extracts data from Spotify API using a search term}
\usage{
search_tracks_by_term(search_term, wasAssociatedWith)
}
\arguments{
\item{search_term}{a string containing the term to search for on Spotify.}

\item{wasAssociatedWith}{an ORCID to write a provenance statement to keep track of who extracted the data set.}
}
\value{
returns a data frame with tracks extracted from Spotify API.
}
\description{
search_tracks_by_term: Extracts data from Spotify API using a search term
}
\examples{
tracks <- search_tracks_by_term("Jana", "https://orcid.org/0000-0003-4253-1953")


}
