% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_track_attributes.R
\docType{data}
\name{track_attributes}
\alias{track_attributes}
\title{Track attributes}
\format{
## `track_attributes`
A list of 17:
\describe{
  \item{isrc}{Track's isrc.}
  \item{id}{Spotify's id for tracks.}
  \item{name}{The name of the track on spotify.}
  \item{duration_ms}{The duration of the recording in ms.}
  \item{album_name}{The name of the album.}
  \item{album_uri}{Spotify album's uri.}
  \item{release_date}{The release date of the album.}
  \item{artist}{The name of the artists on spotify.}
  \item{artist_uri}{Spotify artists' uris.}
}
}
\usage{
track_attributes
}
\description{
A data frame with 10 rows and 9 columns. Each row represents one track and each column and attribute.
}
\details{
This is an example of the attributes that we could possibly keep in our data base.
}
\keyword{datasets}
