% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_identifiers.R
\name{track_identitfiers}
\alias{track_identitfiers}
\title{Extract data that is required for the identification of tracks}
\usage{
track_identitfiers(track)
}
\arguments{
\item{track}{Must be a data frame that contains track information, not artists,
playlists or shows.}

\item{df}{A data frame that may contain the \code{external_ids} column.}
}
\value{
The external_ids column.
}
\description{
Extract data that is required for the identification of tracks
}
