% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xsd_time}
\alias{xsd_time}
\title{XSD time}
\usage{
xsd_time()
}
\value{
A character string with an XSD symbol representation of time.
}
\description{
Create a character string with an XSD symbol representation of time.
}
