% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{provenance_statement}
\alias{provenance_statement}
\title{Create a provenance statement}
\usage{
provenance_statement(
  startedAtTime,
  endedAtTime,
  wasAssociatedWith = "<https://orcid.org/0000-0001-7513-6760>",
  package = "spotifyr"
)
}
\arguments{
\item{startedAtTime}{Time when the collection process was started.}

\item{endedAtTime}{Time when the process ended.}

\item{wasAssociatedWith}{An ORCiD ID, like `<https://orcid.org/0000-0009-9999-8888>`.}
}
\value{
A character string of a provenance statement in Turtle (ttl) format.
}
\description{
Create a provenance statement
}
\examples{
# example code
start_time <- Sys.time()
Sys.sleep(1.2)
end_time <- Sys.time()
provenance_statement (startedAtTime = start_time,
                      endedAtTime = end_time,
                      wasAssociatedWith = '<https://orcid.org/0000-0009-9999-8888>',
                      package = 'spotifyr')
}
