% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R, R/utils.R
\name{xsd_time}
\alias{xsd_time}
\title{xsd_time}
\usage{
xsd_time()

xsd_time()
}
\value{
returns the time stamp of the query.

A character string with an XSD symbol representation of time.
}
\description{
Create a character string with an XSD symbol representation of time.
}
